+++ packages/common/src/types.ts
@@ -1187,16 +1187,16 @@ export interface UpdateCredentialsPayload {
 
 export interface BakerKeysWithProofs {
     signatureVerifyKey: string;
-    signatureKeyProof: string;
     electionVerifyKey: string;
-    electionKeyProof: string;
     aggregationVerifyKey: string;
-    aggregationKeyProof: string;
+    proofAggregation: string;
+    proofSig: string;
+    proofElection: string;
 }
 
 export interface ConfigureBakerPayload {
     /* stake to bake. if set to 0, this removes the account as a baker */
-    stake?: GtuAmount;
+    stake?: CcdAmount;
     /* should earnings from baking be added to staked amount  */
     restakeEarnings?: boolean;
     openForDelegation?: OpenStatus;