+++ packages/rust-bindings/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- `generateBakerKeys`
+
 ## 0.10.0
 
 ### Added